+++ modules/perception/camera/lib/calibrator/common/histogram_estimator.cc
@@ -28,7 +28,7 @@ void HistogramEstimatorParams::Init() {  // set default value
   assert(nr_bins_in_histogram > 0);
   assert(data_ep > data_sp);
 
-  step_bin = (data_ep - data_sp) / nr_bins_in_histogram;
+  step_bin = (data_ep - data_sp) / static_cast<float>(nr_bins_in_histogram);
 
   smooth_kernel.clear();
   smooth_kernel.push_back(1);
@@ -101,7 +101,7 @@ void HistogramEstimator::Smooth(const uint32_t *hist_input, int nr_bins,
 
   float weight_sum_kernel = 0.0f;
   for (auto &w : params_.smooth_kernel) {
-    weight_sum_kernel += w;
+    weight_sum_kernel += static_cast<float>(w);
   }
   float norm_reversed = common::IRec(weight_sum_kernel);
   int sp = 0;
@@ -116,10 +116,10 @@ void HistogramEstimator::Smooth(const uint32_t *hist_input, int nr_bins,
     for (int j = sp; j <= ep; ++j) {
       sum += hist_input[j]
           * params_.smooth_kernel[j - i + filter_radius];
-      kernel_sum += params_.smooth_kernel[j - sp];
+      kernel_sum += static_cast<float>(params_.smooth_kernel[j - sp]);
     }
     float scale = common::IRec(kernel_sum);
-    hist_output[i] = static_cast<uint32_t>(sum * scale);
+    hist_output[i] = static_cast<uint32_t>(static_cast<float>(sum) * scale);
   }
 
   // mid part
@@ -130,7 +130,8 @@ void HistogramEstimator::Smooth(const uint32_t *hist_input, int nr_bins,
     for (int j = sp; j <= ep; ++j) {
       sum += hist_input[j] * params_.smooth_kernel[j - sp];
     }
-    hist_output[i] = static_cast<uint32_t>(sum * norm_reversed);
+    hist_output[i] = static_cast<uint32_t>(static_cast<float>(sum) *
+                                           norm_reversed);
   }
 
   // right boundary part
@@ -141,10 +142,11 @@ void HistogramEstimator::Smooth(const uint32_t *hist_input, int nr_bins,
     float kernel_sum = 0.0f;
     for (int j = sp; j <= ep; ++j) {
       sum += hist_input[j] * params_.smooth_kernel[j - sp];
-      kernel_sum += params_.smooth_kernel[j - sp];
+      kernel_sum += static_cast<float>(params_.smooth_kernel[j - sp]);
     }
     float scale = common::IRec(kernel_sum);
-    hist_output[i] = static_cast<uint32_t>(sum * scale);
+    hist_output[i] = static_cast<uint32_t>(static_cast<float>(sum) *
+                                           scale);
   }
 }
 
@@ -158,8 +160,8 @@ void HistogramEstimator::GenerateHat(float *hist_hat, int nr_bins) {
   float hat_min_allowed = params_.hat_min_allowed;
   int nr_bins_half = nr_bins >> 1;
   for (int i = 0; i < nr_bins; ++i) {
-    float val = (i - nr_bins_half) / hat_std_allowed;
-    hist_hat_[i] = exp(-val * val * 0.5f) + hat_min_allowed;
+    float val = static_cast<float>(i - nr_bins_half) / hat_std_allowed;
+    hist_hat_[i] = expf(-val * val * 0.5f) + hat_min_allowed;
   }
 }
 
@@ -178,7 +180,8 @@ bool HistogramEstimator::IsGoodShape(const uint32_t *hist,
 
   for (int i = sp; i < ep; ++i) {
     float hat_val = hist_hat_[i - offset];
-    if (hist[i] > max_value * hat_val) {
+    if (static_cast<float>(hist[i]) >
+        static_cast<float>(max_value) * hat_val) {
       return false;
     }
   }
